requirements.txt
stickytape
mypy
fake-bpy-module-latest